pyproject.toml
[project]
name = "gpt5-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"pydantic==2.11.7",
	"fastapi==0.116.1",
	"azure-cosmos==4.9.0",
	"uvicorn",
	"httpx",  # for tests
	"pytest"
]

[tool.setuptools]
py-modules = ["main", "models", "db"]
